Update library linking in FindWdk.cmake
The library linking for the target in FindWdk.cmake file has been updated. The BUFFEROVERFLOWK library was removed from the set of libraries linked. The LINK_FLAGS property was also modified to directly specify DriverEntry as the default entry point, and previously related conditional code was commented out for potential future use.